+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/TextUtils.kt
@@ -27,6 +27,8 @@ import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.withStyle
 import androidx.core.util.PatternsCompat
+import io.getstream.chat.android.compose.ui.theme.MentionStyleFactory
+import io.getstream.chat.android.ui.common.feature.messages.composer.mention.Mention
 import java.util.regex.Pattern
 
 internal typealias AnnotationTag = String
@@ -46,6 +48,18 @@ internal const val AnnotationTagEmail: AnnotationTag = "EMAIL"
  */
 internal const val AnnotationTagMention: AnnotationTag = "MENTION"
 
+/**
+ * Builds an [AnnotatedString] from a given text, applying styles and annotations for links and mentions.
+ * Used in message bubbles.
+ *
+ * @param text The input text to be transformed into an [AnnotatedString].
+ * @param textColor The color to be applied to the regular text.
+ * @param textFontStyle The font style to be applied to the regular text.
+ * @param linkStyle The text style to be applied to links within the text.
+ * @param mentionsColor The color to be applied to mentions within the text.
+ * @param mentionedUserNames A list of usernames that are mentioned in the text.
+ * @param builder An optional lambda to apply additional styles or annotations.
+ */
 @SuppressLint("RestrictedApi")
 internal fun buildAnnotatedMessageText(
     text: String,
@@ -89,6 +103,68 @@ internal fun buildAnnotatedMessageText(
             text = text,
             mentionsColor = mentionsColor,
             mentionedUserNames = mentionedUserNames,
+        )
+
+        // Finally, we apply any additional styling that was passed in.
+        builder(this)
+    }
+}
+
+/**
+ * Builds an [AnnotatedString] from a given text, applying styles and annotations for links and mentions.
+ * Used in message input fields.
+ *
+ * @param text The input text to be transformed into an [AnnotatedString].
+ * @param textColor The color to be applied to the regular text.
+ * @param textFontStyle The font style to be applied to the regular text.
+ * @param linkStyle The text style to be applied to links within the text.
+ * @param mentions A set of [Mention] objects representing the mentions in the text.
+ * @param mentionStyleFactory A factory to provide styles for mentions.
+ * @param builder An optional lambda to apply additional styles or annotations.
+ */
+@SuppressLint("RestrictedApi")
+internal fun buildAnnotatedInputText(
+    text: String,
+    textColor: Color,
+    textFontStyle: FontStyle?,
+    linkStyle: TextStyle,
+    mentions: Set<Mention> = emptySet(),
+    mentionStyleFactory: MentionStyleFactory = MentionStyleFactory.NoStyle,
+    builder: (AnnotatedString.Builder).() -> Unit = {},
+): AnnotatedString {
+    return buildAnnotatedString {
+        // First we add the whole text to the [AnnotatedString] and style it as a regular text.
+        append(text)
+        addStyle(
+            SpanStyle(
+                fontStyle = textFontStyle,
+                color = textColor,
+            ),
+            start = 0,
+            end = text.length,
+        )
+
+        // Then for each available link in the text, we add a different style, to represent the links,
+        // as well as add a String annotation to it. This gives us the ability to open the URL on click.
+        linkify(
+            text = text,
+            tag = AnnotationTagUrl,
+            pattern = PatternsCompat.AUTOLINK_WEB_URL,
+            matchFilter = Linkify.sUrlMatchFilter,
+            schemes = URL_SCHEMES,
+            textStyle = linkStyle,
+        )
+        linkify(
+            text = text,
+            tag = AnnotationTagEmail,
+            pattern = PatternsCompat.AUTOLINK_EMAIL_ADDRESS,
+            schemes = EMAIL_SCHEMES,
+            textStyle = linkStyle,
+        )
+        tagMentions(
+            text = text,
+            mentions = mentions,
+            mentionStyleFactory = mentionStyleFactory,
         )
 
         // Finally, we apply any additional styling that was passed in.
@@ -180,6 +256,24 @@ private fun AnnotatedString.Builder.tagUser(
             start = start - 1, // -1 to include the @ symbol
             end = end,
         )
+    }
+}
+
+private fun AnnotatedString.Builder.tagMentions(
+    text: String,
+    mentions: Set<Mention>,
+    mentionStyleFactory: MentionStyleFactory,
+) {
+    mentions.forEach { mention ->
+        val start = text.indexOf(mention.display)
+        val end = start + mention.display.length
+        if (start < 0) return@forEach
+
+        val style = mentionStyleFactory.styleFor(mention)
+        if (style != null) {
+            addStyle(style, start - 1, end) // -1 to include the @ symbol
+            addStringAnnotation(AnnotationTagMention, mention.display, start - 1, end) // -1 to include the @ symbol
+        }
     }
 }